feat: updates backend processing to use cloud-functions-framework
