+++ background-processing/backend/composer.json
@@ -2,11 +2,6 @@
   "require": {
     "google/cloud-translate": "^1.4",
     "google/cloud-firestore": "^1.7",
-    "symfony/http-foundation": "^5.0.0"
-  },
-  "autoload": {
-    "files": [
-      "functions.php"
-    ]
+    "google/cloud-functions-framework": "^0.4.0"
   }
 }
