+++ background-processing/test/backend/translateTest.php
@@ -18,7 +18,7 @@
 use Google\Cloud\Firestore\FirestoreClient;
 use PHPUnit\Framework\TestCase;
 
-require_once __DIR__ . '/../../backend/functions.php';
+require_once __DIR__ . '/../../backend/index.php';
 
 class translateTest extends TestCase
 {
